+++ telers/src/methods/create_forum_topic.rs
@@ -13,7 +13,7 @@ pub struct CreateForumTopic {
     pub name: Box<str>,
     /// Color of the topic icon in RGB format. Currently, must be one of 7322096 (0x6FB9F0), 16766590 (0xFFD67E), 13338331 (0xCB86DB), 9367192 (0x8EEE98), 16749490 (0xFF93B2), or 16478047 (0xFB6F5F).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub icon_color: Option<i64>,
+    pub icon_color: Option<i32>,
     /// Unique identifier of the custom emoji shown as the topic icon. Use [`crate::methods::GetForumTopicIconStickers`] to get all allowed custom emoji identifiers.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub icon_custom_emoji_id: Option<Box<str>>,
@@ -56,14 +56,14 @@ impl CreateForumTopic {
 
     /// Color of the topic icon in RGB format. Currently, must be one of 7322096 (0x6FB9F0), 16766590 (0xFFD67E), 13338331 (0xCB86DB), 9367192 (0x8EEE98), 16749490 (0xFF93B2), or 16478047 (0xFB6F5F).
     #[must_use]
-    pub fn icon_color<T: Into<i64>>(mut self, val: T) -> Self {
+    pub fn icon_color<T: Into<i32>>(mut self, val: T) -> Self {
         self.icon_color = Some(val.into());
         self
     }
 
     /// Color of the topic icon in RGB format. Currently, must be one of 7322096 (0x6FB9F0), 16766590 (0xFFD67E), 13338331 (0xCB86DB), 9367192 (0x8EEE98), 16749490 (0xFF93B2), or 16478047 (0xFB6F5F).
     #[must_use]
-    pub fn icon_color_option<T: Into<i64>>(mut self, val: Option<T>) -> Self {
+    pub fn icon_color_option<T: Into<i32>>(mut self, val: Option<T>) -> Self {
         self.icon_color = val.map(Into::into);
         self
     }
